Removed unused Subject and related classes, exceptions, and tests from Authorization module. Introduced Permission as a value object to replace the removed entity and updated relevant references.
