+++ src/User/Authorization/Domain/Collection/PermissionCollection.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace Webify\User\Authorization\Domain\Collection;
 
 use Webify\Base\Domain\Collection\Collection;
-use Webify\User\Authorization\Domain\Entity\Permission;
+use Webify\User\Authorization\Domain\ValueObject\Permission;
 
 /**
  * Collection class for Permission entities.
